Consistently use int64_t for GBW/GBH/x/y/padded_width/x_minor/minor_width.
This silences several VS 2005 compilation warnings about int64_t being
converted to uint32_t/int and possible loss of data. The warnings were
unintentionally introduced when GBW/GBH/x/y started using int64_t,

Also propagate int64_t use to copy_prev_row().

+++ jbig2dec/jbig2_generic.c
@@ -391,7 +391,7 @@ jbig2_decode_generic_template0(Jbig2Ctx *ctx,
         uint32_t CONTEXT;
         uint32_t line_m1;
         uint32_t line_m2;
-        uint32_t padded_width = (GBW + 7) & -8;
+        int64_t padded_width = (GBW + 7) & -8;
 
         line_m1 = line1 ? line1[0] : 0;
         line_m2 = line2 ? line2[0] << 6 : 0;
@@ -400,8 +400,8 @@ jbig2_decode_generic_template0(Jbig2Ctx *ctx,
         /* 6.2.5.7 3d */
         for (x = 0; x < padded_width; x += 8) {
             byte result = 0;
-            int x_minor;
-            int minor_width = GBW - x > 8 ? 8 : GBW - x;
+            int64_t x_minor;
+            int64_t minor_width = GBW - x > 8 ? 8 : GBW - x;
 
             if (line1)
                 line_m1 = (line_m1 << 8) | (x + 8 < GBW ? line1[(x >> 3) + 1] : 0);
@@ -604,7 +604,7 @@ jbig2_decode_generic_template1(Jbig2Ctx *ctx,
         uint32_t CONTEXT;
         uint32_t line_m1;
         uint32_t line_m2;
-        uint32_t padded_width = (GBW + 7) & -8;
+        int64_t padded_width = (GBW + 7) & -8;
 
         line_m1 = line1 ? line1[0] : 0;
         line_m2 = line2 ? line2[0] << 5 : 0;
@@ -613,8 +613,8 @@ jbig2_decode_generic_template1(Jbig2Ctx *ctx,
         /* 6.2.5.7 3d */
         for (x = 0; x < padded_width; x += 8) {
             byte result = 0;
-            int x_minor;
-            int minor_width = GBW - x > 8 ? 8 : GBW - x;
+            int64_t x_minor;
+            int64_t minor_width = GBW - x > 8 ? 8 : GBW - x;
 
             if (line1)
                 line_m1 = (line_m1 << 8) | (x + 8 < GBW ? line1[(x >> 3) + 1] : 0);
@@ -738,7 +738,7 @@ jbig2_decode_generic_template2(Jbig2Ctx *ctx,
         uint32_t CONTEXT;
         uint32_t line_m1;
         uint32_t line_m2;
-        uint32_t padded_width = (GBW + 7) & -8;
+        int64_t padded_width = (GBW + 7) & -8;
 
         line_m1 = line1 ? line1[0] : 0;
         line_m2 = line2 ? line2[0] << 4 : 0;
@@ -747,8 +747,8 @@ jbig2_decode_generic_template2(Jbig2Ctx *ctx,
         /* 6.2.5.7 3d */
         for (x = 0; x < padded_width; x += 8) {
             byte result = 0;
-            int x_minor;
-            int minor_width = GBW - x > 8 ? 8 : GBW - x;
+            int64_t x_minor;
+            int64_t minor_width = GBW - x > 8 ? 8 : GBW - x;
 
             if (line1)
                 line_m1 = (line_m1 << 8) | (x + 8 < GBW ? line1[(x >> 3) + 1] : 0);
@@ -803,7 +803,7 @@ jbig2_decode_generic_template3(Jbig2Ctx *ctx,
     for (y = 0; y < GBH; y++) {
         uint32_t CONTEXT;
         uint32_t line_m1;
-        uint32_t padded_width = (GBW + 7) & -8;
+        int64_t padded_width = (GBW + 7) & -8;
 
         line_m1 = line1 ? line1[0] : 0;
         CONTEXT = (line_m1 >> 1) & 0x3f0;
@@ -811,8 +811,8 @@ jbig2_decode_generic_template3(Jbig2Ctx *ctx,
         /* 6.2.5.7 3d */
         for (x = 0; x < padded_width; x += 8) {
             byte result = 0;
-            int x_minor;
-            int minor_width = GBW - x > 8 ? 8 : GBW - x;
+            int64_t x_minor;
+            int64_t minor_width = GBW - x > 8 ? 8 : GBW - x;
 
             if (line1)
                 line_m1 = (line_m1 << 8) | (x + 8 < GBW ? line1[(x >> 3) + 1] : 0);
@@ -894,7 +894,7 @@ jbig2_decode_generic_template3_unopt(Jbig2Ctx *ctx,
 }
 
 static void
-copy_prev_row(Jbig2Image *image, int row)
+copy_prev_row(Jbig2Image *image, int64_t row)
 {
     if (!row) {
         /* no previous row */
@@ -918,7 +918,7 @@ jbig2_decode_generic_template0_TPGDON(Jbig2Ctx *ctx,
     int64_t x, y;
     int LTP = 0;
     int gmin, gmax;
-    uint32_t left, right, top;
+    int64_t left, right, top;
 
     if (pixel_outside_field(params->gbat[0], params->gbat[1]) ||
         pixel_outside_field(params->gbat[2], params->gbat[3]) ||
@@ -1011,9 +1011,9 @@ jbig2_decode_generic_template0_TPGDON(Jbig2Ctx *ctx,
         gmin = params->gbat[6];
     if (gmax < params->gbat[6])
         gmax = params->gbat[6];
-    if ((int)left < -gmin)
-        left = -gmin;
-    if ((int)right < gmax)
+    if (left < -((int64_t) gmin))
+        left = -((int64_t) gmin);
+    if (right < gmax)
         right = gmax;
     if (right > GBW)
         right = GBW;
@@ -1030,8 +1030,8 @@ jbig2_decode_generic_template0_TPGDON(Jbig2Ctx *ctx,
         gmin = params->gbat[5];
     if (params->gbat[7] < gmin)
         gmin = params->gbat[7];
-    if ((int)top < -gmin)
-        top = -gmin;
+    if (top < -((int64_t) gmin))
+        top = -((int64_t) gmin);
     /* So 0 <= y < top needs bounds checking. */
 
     for (y = 0; y < GBH; y++) {
